requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.38.6

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.17.2

# Configuration & Environment
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
cryptography==43.0.3

# File Upload & Forms
python-multipart==0.0.9

# Email
fastapi-mail==1.4.1

# Caching
redis==5.0.8

# Storage
supabase==2.10.0
cloudinary==1.41.0

# HTTP Client
httpx==0.27.2

# Google OAuth
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utilities
python-dateutil==2.9.0
typing-extensions==4.12.2
